requirements.txt
strawberry-graphql>=0.200.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dateutil>=2.8.2
Faker>=19.0.0